Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 13ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240118/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.cpp
@@ -47,7 +47,7 @@ int main()
 	long long minCnt = (long long)10e18;
 	long long maxCnt = 0;
 
-	for (int i = m; i <= l; ++i)
+	for (long long i = m; i <= l; ++i)
 	{
 		long long cnt = F(i);
 		minCnt = min(minCnt, cnt);
